+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+- Add: Introduce `TryMigrate` derive macro, this is preferred over the declarative macros. (https://github.com/schneems/magic_migrate/pull/17)
+- Add: Introduce `magic_migrate::MigrateError` as an available generic error for receiving any `TryFrom::Error`. (https://github.com/schneems/magic_migrate/pull/17)
+- Add: Declarative macros (`try_migrate_toml_chain!` etc.) are deprecated in favor of the new derive macro. (https://github.com/schneems/magic_migrate/pull/17)
+
 ## 1.0.1
 
 - Fix: Macro users no longer need to import `std::convert::Infallible` or `serde::de::Deserializer` (https://github.com/schneems/magic_migrate/pull/14)